% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis.R
\name{.asis}
\alias{.asis}
\title{Quick and dirty render pkgdown asis}
\usage{
.asis(..., replacements, notOlderThan = NULL, cacheRepo = "docs/cache")
}
\description{
This will first run pkgdown::build_site then run rmarkdown::render to
build the html pages. This is only relevant if the rmarkdown header
contains an html output format that is not supported by pkgdown
(e.g., ioslides, slidy). This will not work for pdf.
}
\keyword{internal}
