% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis.R
\name{build_articlesCached}
\alias{build_articlesCached}
\title{build_articles from pkgdown, but with Caching}
\usage{
build_articlesCached(pkg = ".", path = "docs/articles", depth = 1L,
  encoding = "UTF-8", quiet = TRUE, notOlderThan = NULL,
  cacheRepo = "cache")
}
\description{
build_articles from pkgdown, but with Caching
}
